added swiping tab layout to the home activity
Added a toolbar and three tabs, one for upcoming events, one for past events and the final one for the user's profile. Tabs can be navigated to either by swiping left or right, or pressing their respective tab title. Note this would not have taken as long to develop had Android updated their documentation on which implementations are deprecated.