fixed permissions and settings checking and added storing fetched GPS coordinates to file
The user is now prompted if either permissions are not assigned or settings are not correct, if the user does not allow permission or change settings, then the user is brought back to the display activity for that event. Note if the user selects to stop showing dialog for requesting permissions, then the record activity cannot start, this will need to be fixed later.
Also added reading the GPS coordinates from file to check that they are correct, displaying them in a second text view component.